requirements.txt
fastapi>=0.104.0  # Updated to support pydantic v2
uvicorn>=0.21.1
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=2.5.2
jinja2>=3.1.2
langsmith>=0.0.63  # Updated to be compatible with langchain
openai>=1.6.1  # Updated to be compatible with langchain-openai
langchain>=0.0.335
langchain-core>=0.1.17
langchain-community>=0.0.1  # Ensure consistent naming with hyphen
requests>=2.31.0
langchain-deepseek>=0.0.2  # Added for DeepSeek integration
redis>=5.0.1

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
requests-mock>=1.11.0

# Optional dependencies - uncomment if needed
# langchain-openai>=0.0.2
langgraph>=0.0.1
